Move vsphere host type into the provider dir
