+++ packages/types/src/types/provider/vsphere/host/VSphereHostInventory.ts
