+++ packages/types/src/types/provider/vsphere/host/index.ts
